+++ App-Project/src/main/java/com/springbootcamp/springsecurity/dto/SellerDto.java
@@ -10,7 +10,7 @@ public class SellerDto extends UserDto {
     @NotEmpty
     private String companyContact;
 
-    public SellerDto(Integer id, String email, String firstName, String middleName, String lastName,
+    public SellerDto(Long id, String email, String firstName, String middleName, String lastName,
                      String gst, String companyContact, String companyName) {
         this.setId(id);
         this.setEmail(email);
