+++ App-Project/src/main/java/com/springbootcamp/springsecurity/repository/CartRepository.java
@@ -3,5 +3,5 @@ package com.springbootcamp.springsecurity.repository;
 import com.springbootcamp.springsecurity.domain.order.Carts;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CartRepository extends CrudRepository<Carts,Integer>{
+public interface CartRepository extends CrudRepository<Carts,Long>{
 }
