+++ App-Project/src/main/java/com/springbootcamp/springsecurity/repository/CategoryRepository.java
@@ -4,7 +4,7 @@ import com.springbootcamp.springsecurity.domain.product.Categories;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface CategoryRepository extends CrudRepository<Categories,Integer> {
+public interface CategoryRepository extends CrudRepository<Categories,Long> {
 
     Categories findByName(String name);
 }
