+++ App-Project/src/main/java/com/springbootcamp/springsecurity/repository/CustomerRepository.java
@@ -3,6 +3,6 @@ package com.springbootcamp.springsecurity.repository;
 import com.springbootcamp.springsecurity.domain.user.Customers;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CustomerRepository extends CrudRepository<Customers,Integer> {
+public interface CustomerRepository extends CrudRepository<Customers,Long> {
     Customers findByEmail(String email);
 }
