+++ App-Project/src/main/java/com/springbootcamp/springsecurity/repository/ProductVariationRepository.java
@@ -3,5 +3,5 @@ package com.springbootcamp.springsecurity.repository;
 import com.springbootcamp.springsecurity.domain.product.ProductVariation;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductVariationRepository extends CrudRepository<ProductVariation,Integer> {
+public interface ProductVariationRepository extends CrudRepository<ProductVariation,Long> {
 }
