+++ App-Project/src/main/java/com/springbootcamp/springsecurity/repository/UserRepository.java
@@ -5,7 +5,7 @@ import com.springbootcamp.springsecurity.domain.user.Users;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface UserRepository extends CrudRepository<Users,Integer> {
+public interface UserRepository extends CrudRepository<Users,Long> {
 
     Users findByEmail(String email);
